+++ RefactoringsExperiments/src/main/java/org/jetbrains/research/extractMethodExperiments/extractors/FalseRefactoringsExtractor.java
@@ -118,7 +118,7 @@ public class FalseRefactoringsExtractor {
                 }
             }.visit(JavaParser.parse(contents), null);
         } catch (Exception e) {
-            logger.log(Level.ERROR, "Could not parse a java file " + e.getMessage());
+            logger.log(Level.ERROR, "Could not parse a java file " + filePath);
         }
     }
 }
