+++ RefactoringsExperiments/src/test/java/org/jetbrains/research/extractMethodExperiments/models/FragmentTest.java
@@ -2,6 +2,7 @@ package org.jetbrains.research.extractMethodExperiments.models;
 
 import org.junit.jupiter.api.Test;
 
+import static org.jetbrains.research.extractMethodExperiments.utils.CodeFormattingUtils.clearCode;
 import static org.jetbrains.research.extractMethodExperiments.utils.feature.generators.DepthAnalyzer.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -31,8 +32,8 @@ class FragmentTest {
 
     @Test
     void clearCodeTest() {
-        assertEquals(codeToTest1, Method.clearCode(codeToTest1), "valid code should remain valid");
-        assertEquals("return i;", Method.clearCode(codeToTest2), "embedded into braces code be cleared");
+        assertEquals(codeToTest1, clearCode(codeToTest1), "valid code should remain valid");
+        assertEquals("return i;", clearCode(codeToTest2), "embedded into braces code be cleared");
     }
 
     @Test
